+++ bin/ghelper
@@ -17,6 +17,60 @@ DOCUMENTATION
     see Git Helper's GitHub repo: https://github.com/emmasax4/git_helper
 """
 
-commands_from File.expand_path(File.dirname(__FILE__) + '../lib/git_helper/commands')
+# todo: break if args are not passed in
+arg :old_owner
+arg :new_owner
+desc "Update a repository's remote URLs from an old GitHub owner to a new owner."
+command 'change-remote' do |c|
+  c.action do |global_options, options, args|
+    require_relative '../lib/git_helper/change_remote.rb'
+    GitHelper::ChangeRemote.new.change_remote(args[0], args[1])
+  end
+end
+
+# todo: change to be an arg and break if it's not passed in
+desc 'Create a GitLab merge request from the current branch.'
+command 'merge-request' do |c|
+  c.switch [:c, :create], desc: 'Create a new pull request'
+  c.switch [:m, :merge], desc: 'Merge an existing pull request'
+
+  c.action do |global_options, options, args|
+    require_relative '../lib/git_helper/merge_request.rb'
+    options = global_options.merge(options)
+
+    if options[:create]
+      GitHelper::GitLabMergeRequest.new.create
+    elsif options[:merge]
+      GitHelper::GitLabMergeRequest.new.merge
+    end
+  end
+end
+
+arg :new_branch_name, optional: true
+desc 'Create a new branch for features, bug fixes, or experimentation.'
+command 'new-branch' do |c|
+  c.action do |global_options, options, args|
+    require_relative '../lib/git_helper/new_branch.rb'
+    GitHelper::NewBranch.new.create_new_branch(args[0])
+  end
+end
+
+# todo: change to be an arg and break if it's not passed in
+desc 'Create a GitHub pull request from the current branch.'
+command 'pull-request' do |c|
+  c.switch [:c, :create], desc: 'Create a new pull request'
+  c.switch [:m, :merge], desc: 'Merge an existing pull request'
+
+  c.action do |global_options, options, args|
+    require_relative '../lib/git_helper/pull_request.rb'
+    options = global_options.merge(options)
+
+    if options[:create]
+      GitHelper::GitHubPullRequest.new.create
+    elsif options[:merge]
+      GitHelper::GitHubPullRequest.new.merge
+    end
+  end
+end
 
 exit run(ARGV)
