+++ lib/git_helper.rb
@@ -3,9 +3,5 @@ require 'yaml'
 require 'gitlab'
 require 'octokit'
 require_relative 'git_helper/version'
-require_relative 'git_helper/commands/change_remote'
-require_relative 'git_helper/commands/merge_request'
-require_relative 'git_helper/commands/new_branch'
-require_relative 'git_helper/commands/pull_request'
 
 module GitHelper; end
