+++ lib/git_helper/git_config_reader.rb
@@ -1,14 +1,12 @@
 require 'yaml'
-require 'octokit'
 
 module GitHelper
-  class OctokitClient
-    # You'll want Octokit version >= 4.18.0, but in your local directory
-    def client
-      Octokit::Client.new(access_token: github_token)
+  class GitConfigReader
+    def gitlab_token
+      config_file[:gitlab_token]
     end
 
-    private def github_token
+    def github_token
       config_file[:github_token]
     end
