+++ lib/git_helper/highline_cli.rb
