+++ lib/git_helper/octokit_client.rb
@@ -1,0 +1,14 @@
+require 'octokit'
+require_relative './git_config_reader.rb'
+
+module GitHelper
+  class OctokitClient
+    def client
+      Octokit::Client.new(access_token: git_config_reader.github_token)
+    end
+
+    private def git_config_reader
+      @git_config_reader ||= GitHelper::GitConfigReader.new
+    end
+  end
+end
