+++ lib/git_helper/prune-merged-branches
